+++ presentation/src/main/java/com/semicolon/walkhub/viewmodel/profile/setting/AccountInfoViewModel.kt
@@ -85,28 +85,6 @@ class AccountInfoViewModel @Inject constructor(
         }
     }
 
-    fun verifyPassword(
-        password: String
-    ) {
-        viewModelScope.launch {
-            kotlin.runCatching {
-                verifyPasswordUseCase.execute(
-                    VerifyPasswordParam(
-                        password
-                    )
-                )
-            }.onFailure {
-                when (it) {
-                    is BadRequestException -> event(Event.ErrorMessage("현재 비밀번호가 틀렸습니다. 알맞은 비밀번호를 입력해주세요."))
-                    is UnauthorizedException -> event(Event.ErrorMessage("세션이 만료되었습니다. 다시 시도해주세요."))
-                    is NoInternetException -> event(Event.ErrorMessage("인터넷에 연결되어있지 않습니다."))
-                    is NullPointerException -> event(Event.ErrorMessage("현재 비밀번호를 입력해주세요."))
-                    else -> event(Event.ErrorMessage("에러가 발생했습니다."))
-                }
-            }
-        }
-    }
-
 
     private fun FetchAuthInfoEntity.toData() =
         FetchAuthInfoEntity(
